refactored: removed android.support.v13.app.FragmentPagerAdapter and replaced with android.support.v4.app.FragmentPagerAdapter
