+++ WordPress/src/main/java/org/wordpress/android/ui/plans/PlanFragment.java
@@ -1,8 +1,8 @@
 package org.wordpress.android.ui.plans;
 
-import android.app.Fragment;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
+import android.support.v4.app.Fragment;
 import android.text.Html;
 import android.text.TextUtils;
 import android.view.LayoutInflater;
