+++ WordPress/src/main/java/org/wordpress/android/ui/posts/EditPostPreviewFragment.java
@@ -1,10 +1,10 @@
 package org.wordpress.android.ui.posts;
 
-import android.app.Fragment;
 import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
 import android.text.SpannableString;
 import android.text.Spanned;
 import android.view.LayoutInflater;
