+++ WordPress/src/main/java/org/wordpress/android/ui/posts/EditPostSettingsFragment.java
@@ -1,13 +1,13 @@
 package org.wordpress.android.ui.posts;
 
 import android.app.Activity;
-import android.app.Fragment;
 import android.content.Intent;
 import android.location.Address;
 import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.view.menu.MenuPopupHelper;
@@ -428,7 +428,7 @@ public class EditPostSettingsFragment extends Fragment {
                         updateExcerpt(input);
                     }
                 });
-        dialog.show(getFragmentManager(), null);
+        dialog.show(getChildFragmentManager(), null);
     }
 
     private void showSlugDialog() {
