+++ WordPress/src/main/java/org/wordpress/android/ui/reader/ReaderBlogFragment.java
@@ -1,7 +1,7 @@
 package org.wordpress.android.ui.reader;
 
-import android.app.Fragment;
 import android.os.Bundle;
+import android.support.v4.app.Fragment;
 import android.support.v4.view.MenuItemCompat;
 import android.support.v7.widget.SearchView;
 import android.text.TextUtils;
