+++ WordPress/src/main/java/org/wordpress/android/ui/reader/ReaderTagFragment.java
@@ -1,8 +1,8 @@
 package org.wordpress.android.ui.reader;
 
-import android.app.Fragment;
 import android.content.Context;
 import android.os.Bundle;
+import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
